S7: the eval record stops re-narrowing constants to float32
data_sources.py:191 is the only coercion between symbolic-data's Problem and
the eval record, and the srbf-side twin of the generation gate. At float32 it
silently re-narrowed the constants symbolic-data had just stopped narrowing
(S6), so the widening would have died one layer short of the benchmark.

The fixtures that build Problems by hand move with it. NOT touched:
test_reference_metrics.py's float32-eps assertion -- that is a TOLERANCE (the
noise floor a fitted law must clear), not a storage width, and the same goes
for is_perfect_fit's bar in metrics/numeric.py.

421 passed.

Note for whoever hits it next: srbf's own .venv cannot collect 8 test modules
because safetensors is missing -- it arrives transitively via flash-ansr and is
not declared in srbf's pyproject. Pre-existing, verified against a clean tree,
and unrelated to this change; the run above used a complete environment.

+++ src/srbf/data_sources.py
@@ -188,7 +188,10 @@ class CatalogSource(EvaluationDataSource):
         return {
             "input_ids": np.asarray(input_ids, dtype=np.int64) if input_ids is not None else None,
             "labels": np.asarray(input_ids[1:], dtype=np.int64) if input_ids is not None else None,
-            "constants": [np.asarray(problem.constants, dtype=np.float32)] if problem.constants else [],
+            # float64: the only coercion between symbolic-data's Problem and the eval record,
+            # and the srbf-side twin of the generation gate. At float32 it silently
+            # re-narrowed constants the generator had just stopped narrowing.
+            "constants": [np.asarray(problem.constants, dtype=np.float64)] if problem.constants else [],
             "benchmark_eq_id": problem.eq_id,
             "variable_names": list(problem.variables) if problem.variables else None,
             "ground_truth_prefix": gt_prefix,
